Migrate sanitizer BlockDef methods into defs package
